feat(platform): ability to list scene transitions
Provides an API to list scene transitions from platform apps.
Includes a separate method to filter those belonging to the current app. Returned objects are the
bare minimum (see `ITransition`) and *does not* include settings or property manager settings, we
will choose what to expose based on developer usage. Current use case is set default, and we only
need an ID for that.